refactor export utility
